offline_tool/requirements.txt
pandas
polars
numpy
python-can
cantools
pyarrow
openpyxl
pillow
opencv-python
ffmpeg-python
typer
requests
pyyaml
matplotlib
seaborn
plotly
asammdf
scipy
scikit-learn
python-docx
python-pptx
tqdm
PySide6
